Add flag (#32)
* add flags

* add flags

* fixed conflict

* fixed conflict.

* fixed bug

+++ src/cnn/layers.hpp
@@ -53,9 +53,9 @@ void Convolution<w_row, w_col, input, output, P, S, T>
   CnnProto::Weight* w = p->add_weights();
   CnnProto::Bias* b = p->add_biases();
   for (int i = 0; i < w_.size(); ++i)
-    w->mutable_w()->Add(Converter::ToDouble(w_[i]));
+    w->mutable_w()->Add(Converter::ToFloat(w_[i]));
   for (int i = 0; i < b_.size(); ++i)
-    b->mutable_b()->Add(Converter::ToDouble(b_[i]));
+    b->mutable_b()->Add(Converter::ToFloat(b_[i]));
 }
 
 template<int w_row, int w_col, int input, int output, int P, int S, typename T>
@@ -194,9 +194,9 @@ void Affine<input, output, T>::saveParams(CnnProto::Params *p) const {
   CnnProto::Weight* w = p->add_weights();
   CnnProto::Bias* b = p->add_biases();
   for (int i = 0; i < w_.size(); ++i)
-    w->mutable_w()->Add(Converter::ToDouble(w_[i]));
+    w->mutable_w()->Add(Converter::ToFloat(w_[i]));
   for (int i = 0; i < b_.size(); ++i)
-    b->mutable_b()->Add(Converter::ToDouble(b_[i]));
+    b->mutable_b()->Add(Converter::ToFloat(b_[i]));
 }
 
 template<int input, int output, typename T>
